Introduce notify function, and be more verbose
Call notify-send from a function, so that it is easier to notify
multiple places in the script. This will also make it easier to disable
the notifications, or push notifications through another command in the
future.

At the same time, make the script a bit more verbose about what happens,
so that a notification will be presented when the YubiKey should be
touched, and another one, once the code is copied. This makes it easier
to use the script through a keyboard shortcut.